+++ packages/datastore/__tests__/helpers.ts
@@ -1,16 +1,21 @@
+import Observable from 'zen-observable-ts';
+import { ModelInit, Schema, InternalSchema, __modelMeta__ } from '../src/types';
 import {
 	AsyncCollection,
-	ModelInit,
 	MutableModel,
-	Schema,
-	InternalSchema,
+	DataStore as DS,
 	CompositeIdentifier,
 	CustomIdentifier,
-	__modelMeta__,
-	SchemaModel,
+	ManagedIdentifier,
+	PersistentModel,
+	// IdentifierFields,
+	// ModelInit,
+	// PersistentModel,
 	PersistentModelConstructor,
 	OptionallyManagedIdentifier,
-} from '../src/types';
+	// Predicates,
+	// __modelMeta__,
+} from '../src';
 
 import {
 	initSchema as _initSchema,
@@ -64,6 +69,14 @@ export function expectMutation(mutation, values) {
 			`Bad mutation: ${JSON.stringify(data, null, 2)}\n${errors.join('\n')}`
 		);
 	}
+}
+
+export function expectType<T>(_param: T): _param is T {
+	return true;
+}
+
+export function dummyInstance<T extends PersistentModel>(): T {
+	return <T>{};
 }
 
 /**
@@ -437,6 +450,27 @@ export function getDataStore() {
 }
 
 // #region schemas
+export const DataStore: typeof DS = (() => {
+	class clazz {}
+
+	const proxy = new Proxy(clazz, {
+		get: (_, prop) => {
+			const p = prop as keyof typeof DS;
+
+			switch (p) {
+				case 'query':
+				case 'save':
+				case 'delete':
+					return () => new Proxy({}, {});
+				case 'observe':
+				case 'observeQuery':
+					return () => Observable.of();
+			}
+		},
+	}) as unknown as typeof DS;
+
+	return proxy;
+})();
 
 export declare class Model {
 	public readonly id: string;
@@ -698,10 +732,8 @@ export declare class CompositePKParent {
 	readonly content: string;
 	readonly children?: (CompositePKChild | null)[] | null;
 	readonly implicitChildren?: (ImplicitChild | null)[] | null;
-	readonly strangeExplicitChildren?: (StrangeExplicitChild | null)[] | null;
-	readonly childrenWithoutBelongsTo?:
-		| (ChildWithoutReturnBelongsTo | null)[]
-		| null;
+	readonly strangeChildren?: (StrangeExplicitChild | null)[] | null;
+	readonly childrenSansBelongsTo?: (ChildSansBelongsTo | null)[] | null;
 	readonly createdAt?: string | null;
 	readonly updatedAt?: string | null;
 	constructor(init: ModelInit<CompositePKParent>);
@@ -759,11 +791,11 @@ export declare class StrangeExplicitChild {
 	readonly [__modelMeta__]: {
 		identifier: CompositeIdentifier<
 			StrangeExplicitChild,
-			['strangeExplicitChildId', 'content']
+			['strangeId', 'content']
 		>;
 		readOnlyFields: 'createdAt' | 'updatedAt';
 	};
-	readonly strangeExplicitChildId: string;
+	readonly strangeId: string;
 	readonly content: string;
 	readonly parent: CompositePKParent;
 	readonly strangeParentId?: string | null;
@@ -779,27 +811,24 @@ export declare class StrangeExplicitChild {
 	): StrangeExplicitChild;
 }
 
-export declare class ChildWithoutReturnBelongsTo {
+export declare class ChildSansBelongsTo {
 	readonly [__modelMeta__]: {
-		identifier: CompositeIdentifier<
-			ChildWithoutReturnBelongsTo,
-			['childId', 'content']
-		>;
+		identifier: CompositeIdentifier<ChildSansBelongsTo, ['childId', 'content']>;
 		readOnlyFields: 'createdAt' | 'updatedAt';
 	};
 	readonly childId: string;
 	readonly content: string;
+	readonly compositePKParentChildrenSansBelongsToCustomId: string;
+	readonly compositePKParentChildrenSansBelongsToContent?: string | null;
 	readonly createdAt?: string | null;
 	readonly updatedAt?: string | null;
-	readonly compositePKParentChildrenWithoutBelongsToCustomId?: string | null;
-	readonly compositePKParentChildrenWithoutBelongsToContent?: string | null;
-	constructor(init: ModelInit<ChildWithoutReturnBelongsTo>);
+	constructor(init: ModelInit<ChildSansBelongsTo>);
 	static copyOf(
-		source: ChildWithoutReturnBelongsTo,
+		source: ChildSansBelongsTo,
 		mutator: (
-			draft: MutableModel<ChildWithoutReturnBelongsTo>
-		) => MutableModel<ChildWithoutReturnBelongsTo> | void
-	): ChildWithoutReturnBelongsTo;
+			draft: MutableModel<ChildSansBelongsTo>
+		) => MutableModel<ChildSansBelongsTo> | void
+	): ChildSansBelongsTo;
 }
 
 export function testSchema(): Schema {
@@ -1644,8 +1673,8 @@ export function testSchema(): Schema {
 							],
 						},
 					},
-					strangeExplicitChildren: {
-						name: 'strangeExplicitChildren',
+					strangeChildren: {
+						name: 'strangeChildren',
 						isArray: true,
 						type: {
 							model: 'StrangeExplicitChild',
@@ -1658,11 +1687,11 @@ export function testSchema(): Schema {
 							associatedWith: ['parent'],
 						},
 					},
-					childrenWithoutBelongsTo: {
-						name: 'childrenWithoutBelongsTo',
+					childrenSansBelongsTo: {
+						name: 'childrenSansBelongsTo',
 						isArray: true,
 						type: {
-							model: 'ChildWithoutReturnBelongsTo',
+							model: 'ChildSansBelongsTo',
 						},
 						isRequired: false,
 						attributes: [],
@@ -1670,8 +1699,8 @@ export function testSchema(): Schema {
 						association: {
 							connectionType: 'HAS_MANY',
 							associatedWith: [
-								'compositePKParentChildrenWithoutBelongsToCustomId',
-								'compositePKParentChildrenWithoutBelongsToContent',
+								'compositePKParentChildrenSansBelongsToCustomId',
+								'compositePKParentChildrenSansBelongsToContent',
 							],
 						},
 					},
@@ -1703,12 +1732,6 @@ export function testSchema(): Schema {
 						type: 'key',
 						properties: {
 							fields: ['customId', 'content'],
-						},
-					},
-					{
-						type: 'key',
-						properties: {
-							fields: ['content'],
 						},
 					},
 				],
@@ -1819,7 +1842,7 @@ export function testSchema(): Schema {
 						type: {
 							model: 'CompositePKParent',
 						},
-						isRequired: false,
+						isRequired: true,
 						attributes: [],
 						association: {
 							connectionType: 'BELONGS_TO',
@@ -1878,8 +1901,8 @@ export function testSchema(): Schema {
 			StrangeExplicitChild: {
 				name: 'StrangeExplicitChild',
 				fields: {
-					strangeExplicitChildId: {
-						name: 'strangeExplicitChildId',
+					strangeId: {
+						name: 'strangeId',
 						isArray: false,
 						type: 'ID',
 						isRequired: true,
@@ -1898,7 +1921,7 @@ export function testSchema(): Schema {
 						type: {
 							model: 'CompositePKParent',
 						},
-						isRequired: false,
+						isRequired: true,
 						attributes: [],
 						association: {
 							connectionType: 'BELONGS_TO',
@@ -1946,7 +1969,7 @@ export function testSchema(): Schema {
 					{
 						type: 'key',
 						properties: {
-							fields: ['strangeExplicitChildId', 'content'],
+							fields: ['strangeId', 'content'],
 						},
 					},
 					{
@@ -1958,8 +1981,8 @@ export function testSchema(): Schema {
 					},
 				],
 			},
-			ChildWithoutReturnBelongsTo: {
-				name: 'ChildWithoutReturnBelongsTo',
+			ChildSansBelongsTo: {
+				name: 'ChildSansBelongsTo',
 				fields: {
 					childId: {
 						name: 'childId',
@@ -1973,6 +1996,20 @@ export function testSchema(): Schema {
 						isArray: false,
 						type: 'String',
 						isRequired: true,
+						attributes: [],
+					},
+					compositePKParentChildrenSansBelongsToCustomId: {
+						name: 'compositePKParentChildrenSansBelongsToCustomId',
+						isArray: false,
+						type: 'ID',
+						isRequired: true,
+						attributes: [],
+					},
+					compositePKParentChildrenSansBelongsToContent: {
+						name: 'compositePKParentChildrenSansBelongsToContent',
+						isArray: false,
+						type: 'String',
+						isRequired: false,
 						attributes: [],
 					},
 					createdAt: {
@@ -1991,23 +2028,9 @@ export function testSchema(): Schema {
 						attributes: [],
 						isReadOnly: true,
 					},
-					compositePKParentChildrenWithoutBelongsToCustomId: {
-						name: 'compositePKParentChildrenWithoutBelongsToCustomId',
-						isArray: false,
-						type: 'ID',
-						isRequired: false,
-						attributes: [],
-					},
-					compositePKParentChildrenWithoutBelongsToContent: {
-						name: 'compositePKParentChildrenWithoutBelongsToContent',
-						isArray: false,
-						type: 'String',
-						isRequired: false,
-						attributes: [],
-					},
 				},
 				syncable: true,
-				pluralName: 'ChildWithoutReturnBelongsTos',
+				pluralName: 'ChildSansBelongsTos',
 				attributes: [
 					{
 						type: 'model',
@@ -2017,6 +2040,16 @@ export function testSchema(): Schema {
 						type: 'key',
 						properties: {
 							fields: ['childId', 'content'],
+						},
+					},
+					{
+						type: 'key',
+						properties: {
+							name: 'byParent',
+							fields: [
+								'compositePKParentChildrenSansBelongsToCustomId',
+								'compositePKParentChildrenSansBelongsToContent',
+							],
 						},
 					},
 				],
@@ -2389,3 +2422,276 @@ export function smallTestSchema(): Schema {
 }
 
 // #endregion schemas
+
+//#region Types
+
+//#region Legacy
+
+export type LegacyCustomROMETA = {
+	readOnlyFields: 'createdOn' | 'updatedOn';
+};
+
+export class LegacyCustomRO {
+	readonly id: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdOn?: string;
+	readonly updatedOn?: string;
+	constructor(init: ModelInit<LegacyCustomRO, LegacyCustomROMETA>) {}
+	static copyOf(
+		source: LegacyCustomRO,
+		mutator: (
+			draft: MutableModel<LegacyCustomRO, LegacyCustomROMETA>
+		) => MutableModel<LegacyCustomRO, LegacyCustomROMETA> | void
+	): LegacyCustomRO {
+		return <LegacyCustomRO>(<unknown>undefined);
+	}
+}
+
+export type LegacyDefaultROMETA = {
+	readOnlyFields: 'createdAt' | 'updatedAt';
+};
+
+export class LegacyDefaultRO {
+	readonly id: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdAt?: string;
+	readonly updatedAt?: string;
+	constructor(init: ModelInit<LegacyDefaultRO, LegacyDefaultROMETA>) {}
+	static copyOf(
+		source: LegacyDefaultRO,
+		mutator: (
+			draft: MutableModel<LegacyDefaultRO, LegacyDefaultROMETA>
+		) => MutableModel<LegacyDefaultRO, LegacyDefaultROMETA> | void
+	): LegacyDefaultRO {
+		return <LegacyDefaultRO>(<unknown>undefined);
+	}
+}
+
+export class LegacyNoMetadata {
+	readonly id: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdAt?: string;
+	readonly updatedAt?: string;
+	constructor(init: ModelInit<LegacyNoMetadata>) {}
+	static copyOf(
+		source: LegacyNoMetadata,
+		mutator: (
+			draft: MutableModel<LegacyNoMetadata>
+		) => MutableModel<LegacyNoMetadata> | void
+	): LegacyNoMetadata {
+		return <LegacyNoMetadata>(<unknown>undefined);
+	}
+}
+
+//#endregion
+
+//#region Managed
+
+export class ManagedCustomRO {
+	readonly [__modelMeta__]: {
+		identifier: ManagedIdentifier<ManagedCustomRO, 'id'>;
+		readOnlyFields: 'createdOn' | 'updatedOn';
+	};
+	readonly id: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdOn?: string;
+	readonly updatedOn?: string;
+	constructor(init: ModelInit<ManagedCustomRO>) {}
+	static copyOf(
+		source: ManagedCustomRO,
+		mutator: (
+			draft: MutableModel<ManagedCustomRO>
+		) => MutableModel<ManagedCustomRO> | void
+	): ManagedCustomRO {
+		return <ManagedCustomRO>(<unknown>undefined);
+	}
+}
+
+export class ManagedDefaultRO {
+	readonly [__modelMeta__]: {
+		identifier: ManagedIdentifier<ManagedDefaultRO, 'id'>;
+		readOnlyFields: 'createdAt' | 'updatedAt';
+	};
+	readonly id: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdAt?: string;
+	readonly updatedAt?: string;
+	constructor(init: ModelInit<ManagedDefaultRO>) {}
+	static copyOf(
+		source: ManagedDefaultRO,
+		mutator: (
+			draft: MutableModel<ManagedDefaultRO>
+		) => MutableModel<ManagedDefaultRO> | void
+	): ManagedDefaultRO {
+		return <ManagedDefaultRO>(<unknown>undefined);
+	}
+}
+
+//#endregion
+
+//#region Optionally Managed
+
+export class OptionallyManagedCustomRO {
+	readonly [__modelMeta__]: {
+		identifier: OptionallyManagedIdentifier<OptionallyManagedCustomRO, 'id'>;
+		readOnlyFields: 'createdOn' | 'updatedOn';
+	};
+	readonly id: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdOn?: string;
+	readonly updatedOn?: string;
+	constructor(init: ModelInit<OptionallyManagedCustomRO>) {}
+	static copyOf(
+		source: OptionallyManagedCustomRO,
+		mutator: (
+			draft: MutableModel<OptionallyManagedCustomRO>
+		) => MutableModel<OptionallyManagedCustomRO> | void
+	): OptionallyManagedCustomRO {
+		return <OptionallyManagedCustomRO>(<unknown>undefined);
+	}
+}
+
+export class OptionallyManagedDefaultRO {
+	readonly [__modelMeta__]: {
+		identifier: OptionallyManagedIdentifier<OptionallyManagedDefaultRO, 'id'>;
+		readOnlyFields: 'createdAt' | 'updatedAt';
+	};
+	readonly id: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdAt?: string;
+	readonly updatedAt?: string;
+	constructor(init: ModelInit<OptionallyManagedDefaultRO>) {}
+	static copyOf(
+		source: OptionallyManagedDefaultRO,
+		mutator: (
+			draft: MutableModel<OptionallyManagedDefaultRO>
+		) => MutableModel<OptionallyManagedDefaultRO> | void
+	): OptionallyManagedDefaultRO {
+		return <OptionallyManagedDefaultRO>(<unknown>undefined);
+	}
+}
+
+//#endregion
+
+//#region Composite
+
+export class CompositeCustomRO {
+	readonly [__modelMeta__]: {
+		identifier: CompositeIdentifier<CompositeCustomRO, ['tenant', 'dob']>;
+		readOnlyFields: 'createdOn' | 'updatedOn';
+	};
+	readonly tenant: string;
+	readonly dob: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdOn?: string;
+	readonly updatedOn?: string;
+	constructor(init: ModelInit<CompositeCustomRO>) {}
+	static copyOf(
+		source: CompositeCustomRO,
+		mutator: (
+			draft: MutableModel<CompositeCustomRO>
+		) => MutableModel<CompositeCustomRO> | void
+	): CompositeCustomRO {
+		return <CompositeCustomRO>(<unknown>undefined);
+	}
+}
+
+export class CompositeDefaultRO {
+	readonly [__modelMeta__]: {
+		identifier: CompositeIdentifier<CompositeDefaultRO, ['tenant', 'dob']>;
+		readOnlyFields: 'createdAt' | 'updatedAt';
+	};
+	readonly tenant: string;
+	readonly dob: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdAt?: string;
+	readonly updatedAt?: string;
+	constructor(init: ModelInit<CompositeDefaultRO>) {}
+	static copyOf(
+		source: CompositeDefaultRO,
+		mutator: (
+			draft: MutableModel<CompositeDefaultRO>
+		) => MutableModel<CompositeDefaultRO> | void
+	): CompositeDefaultRO {
+		return <CompositeDefaultRO>(<unknown>undefined);
+	}
+}
+
+//#endregion
+
+//#region Custom
+
+export class CustomIdentifierCustomRO {
+	readonly [__modelMeta__]: {
+		identifier: CustomIdentifier<CustomIdentifierCustomRO, 'myId'>;
+		readOnlyFields: 'createdOn' | 'updatedOn';
+	};
+	readonly myId: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdOn: string;
+	readonly updatedOn: string;
+	constructor(init: ModelInit<CustomIdentifierCustomRO>) {}
+	static copyOf(
+		source: CustomIdentifierCustomRO,
+		mutator: (
+			draft: MutableModel<CustomIdentifierCustomRO>
+		) => MutableModel<CustomIdentifierCustomRO> | void
+	): CustomIdentifierCustomRO {
+		return <CustomIdentifierCustomRO>(<unknown>undefined);
+	}
+}
+
+export class CustomIdentifierDefaultRO {
+	readonly [__modelMeta__]: {
+		identifier: CustomIdentifier<CustomIdentifierDefaultRO, 'myId'>;
+		readOnlyFields: 'createdAt' | 'updatedAt';
+	};
+	readonly myId: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdAt?: string;
+	readonly updatedAt?: string;
+	constructor(init: ModelInit<CustomIdentifierDefaultRO>) {}
+	static copyOf(
+		source: CustomIdentifierDefaultRO,
+		mutator: (
+			draft: MutableModel<CustomIdentifierDefaultRO>
+		) => MutableModel<CustomIdentifierDefaultRO> | void
+	): CustomIdentifierDefaultRO {
+		return <CustomIdentifierDefaultRO>(<unknown>undefined);
+	}
+}
+
+export class CustomIdentifierNoRO {
+	readonly [__modelMeta__]: {
+		identifier: CustomIdentifier<CustomIdentifierNoRO, 'myId'>;
+	};
+	readonly myId: string;
+	readonly name: string;
+	readonly description?: string;
+	readonly createdAt?: string;
+	readonly updatedAt?: string;
+	constructor(init: ModelInit<CustomIdentifierNoRO>) {}
+	static copyOf(
+		source: CustomIdentifierNoRO,
+		mutator: (
+			draft: MutableModel<CustomIdentifierNoRO>
+		) => MutableModel<CustomIdentifierNoRO> | void
+	): CustomIdentifierDefaultRO {
+		return undefined!;
+	}
+}
+
+//#endregion
+
+//#endregion