+++ packages/interactions/src/types/index.ts
@@ -13,4 +13,5 @@
 export * from './Interactions';
 export * from './Provider';
 export * from './Providers/AWSLexProvider';
+export * from './Providers/AWSLexV2Provider';
 export * from './Response';