+++ packages/amplify-ui-angular/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.62](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-angular@1.0.61...@aws-amplify/ui-angular@1.0.62) (2022-09-08)
+
+**Note:** Version bump only for package @aws-amplify/ui-angular
+
+## [1.0.61](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-angular@1.0.60...@aws-amplify/ui-angular@1.0.61) (2022-09-01)
+
+**Note:** Version bump only for package @aws-amplify/ui-angular
+
 ## [1.0.60](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-angular@1.0.59...@aws-amplify/ui-angular@1.0.60) (2022-08-23)
 
 **Note:** Version bump only for package @aws-amplify/ui-angular